+++ sdk/communication/Azure.Communication.Sms/src/Models/SendMessageRequest.cs
@@ -3,8 +3,10 @@
 
 using Azure.Core;
 
-namespace Azure.Communication.Sms
+namespace Azure.Communication.Sms.Models
 {
+    /// <summary> Send Message Request. </summary>
+
     [CodeGenModel("SendMessageRequest")]
     internal partial class SendMessageRequest
     {