+++ sdk/communication/Azure.Communication.Sms/src/SmsClient.cs
@@ -3,12 +3,14 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core;
 using Azure.Core.Pipeline;
 using Azure.Communication.Pipeline;
+using System.Linq;
+using Azure.Communication.Sms.Models;
+using System.Globalization;
 
 namespace Azure.Communication.Sms
 {
@@ -98,17 +100,19 @@ namespace Azure.Communication.Sms
         /// <param name="from">The sender's phone number that is owned by the authenticated account.</param>
         /// <param name="to">The recipient's phone number.</param>
         /// <param name="message">The contents of the message that will be sent to the recipient. The allowable content is defined by RFC 5724. If the message has more than 160 characters, the server will split it into multiple SMSs automatically.</param>
-        /// <param name="sendSmsOptions">Optional configuration for sending SMS messages.</param>
+        /// <param name="options">Optional configuration for sending SMS messages.</param>
         /// <param name="cancellationToken">The cancellation token for the task.</param>
         /// <exception cref="RequestFailedException">The server returned an error. See <see cref="Exception.Message"/> for details returned from the server.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="from"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="to"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="message"/> is null.</exception>
-        public virtual async Task<Response<SendSmsResponse>> SendAsync(PhoneNumberIdentifier from, PhoneNumberIdentifier to, string message, SendSmsOptions sendSmsOptions = null, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<SmsSendResult>> SendAsync(string from, string to, string message, SmsSendOptions options = default, CancellationToken cancellationToken = default)
         {
-            Argument.AssertNotNullOrEmpty(from.PhoneNumber, nameof(from));
-            Argument.AssertNotNullOrEmpty(to.PhoneNumber, nameof(to));
-            return await SendAsync(from, new[] { to }, message, sendSmsOptions, cancellationToken).ConfigureAwait(false);
+            Argument.AssertNotNullOrEmpty(from, nameof(from));
+            Argument.AssertNotNullOrEmpty(to, nameof(to));
+            Response<IEnumerable<SmsSendResult>> allSmsSendResults = await SendAsync(from, new[] { to }, message, options, cancellationToken).ConfigureAwait(false);
+            SmsSendResult sendSMSResult = allSmsSendResults.Value.FirstOrDefault();
+            return Response.FromValue(sendSMSResult, allSmsSendResults.GetRawResponse());
         }
 
         /// <summary>
@@ -117,37 +121,50 @@ namespace Azure.Communication.Sms
         /// <param name="from">The sender's phone number that is owned by the authenticated account.</param>
         /// <param name="to">The recipient's phone number.</param>
         /// <param name="message">The contents of the message that will be sent to the recipient. The allowable content is defined by RFC 5724. If the message has more than 160 characters, the server will split it into multiple SMSs automatically.</param>
-        /// <param name="sendSmsOptions">Optional configuration for sending SMS messages.</param>
+        /// <param name="options">Optional configuration for sending SMS messages.</param>
         /// <param name="cancellationToken">The cancellation token for the underlying request.</param>
         /// <exception cref="RequestFailedException">The server returned an error. See <see cref="Exception.Message"/> for details returned from the server.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="from"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="to"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="message"/> is null.</exception>
-        public virtual Response<SendSmsResponse> Send(PhoneNumberIdentifier from, PhoneNumberIdentifier to, string message, SendSmsOptions sendSmsOptions = null, CancellationToken cancellationToken = default)
+        public virtual Response<SmsSendResult> Send(string from, string to, string message, SmsSendOptions options = default, CancellationToken cancellationToken = default)
         {
-            Argument.AssertNotNullOrEmpty(from.PhoneNumber, nameof(from));
-            Argument.AssertNotNullOrEmpty(to.PhoneNumber, nameof(to));
-            return Send(from, new[] { to }, message, sendSmsOptions, cancellationToken);
+            Argument.AssertNotNullOrEmpty(from, nameof(from));
+            Argument.AssertNotNullOrEmpty(to, nameof(to));
+            Response<IEnumerable<SmsSendResult>> allSmsSendResults = Send(from, new[] { to }, message, options, cancellationToken);
+            SmsSendResult sendSMSResult = allSmsSendResults.Value.FirstOrDefault();
+            return Response.FromValue(sendSMSResult, allSmsSendResults.GetRawResponse());
         }
 
         /// <summary> Sends an SMS message from a phone number that belongs to the authenticated account. </summary>
         /// <param name="from"> The sender&apos;s phone number in E.164 format that is owned by the authenticated account. </param>
-        /// <param name="to"> The recipient&apos;s phone number in E.164 format. In this version, only one recipient in the list is supported. </param>
+        /// <param name="to"> The recipient&apos;s phone number in E.164 format. In this version, up to 100 recipients in the list is supported. </param>
         /// <param name="message"> The contents of the message that will be sent to the recipient. The allowable content is defined by RFC 5724. </param>
-        /// <param name="sendSmsOptions"> Optional configuration for sending SMS messages. </param>
+        /// <param name="options"> Optional configuration for sending SMS messages. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="RequestFailedException">The server returned an error. See <see cref="Exception.Message"/> for details returned from the server.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="from"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="to"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="message"/> is null.</exception>
-        public virtual async Task<Response<SendSmsResponse>> SendAsync(PhoneNumberIdentifier from, IEnumerable<PhoneNumberIdentifier> to, string message, SendSmsOptions sendSmsOptions = null, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<IEnumerable<SmsSendResult>>> SendAsync(string from, IEnumerable<string> to, string message, SmsSendOptions options = default, CancellationToken cancellationToken = default)
         {
             using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(SmsClient)}.{nameof(Send)}");
             scope.Start();
             try
             {
-                Argument.AssertNotNullOrEmpty(from.PhoneNumber, nameof(from));
-                return await RestClient.SendAsync(from.PhoneNumber, to.Select(x => AssertNotNullOrEmpty(x.PhoneNumber, nameof(to))), message, sendSmsOptions, cancellationToken).ConfigureAwait(false);
+                Argument.AssertNotNullOrEmpty(from, nameof(from));
+                List<SmsRecipient> smsRecipients = new List<SmsRecipient>();
+                foreach (string toPhoneNumber in to)
+                {
+                    Argument.AssertNotNullOrEmpty(toPhoneNumber, nameof(to));
+                    SmsRecipient smsRecipient = new SmsRecipient(toPhoneNumber);
+                    smsRecipient.RepeatabilityRequestId = Guid.NewGuid().ToString();
+                    smsRecipient.RepeatabilityFirstSent = DateTime.Now.ToString(CultureInfo.InvariantCulture);
+                    smsRecipients.Add(smsRecipient);
+                }
+
+                Response<SmsSendResponse> response = await RestClient.SendAsync(from, smsRecipients, message, options, cancellationToken).ConfigureAwait(false);
+                return Response.FromValue((IEnumerable<SmsSendResult>)response.Value.Value, response.GetRawResponse());
             }
             catch (Exception ex)
             {
@@ -158,22 +175,33 @@ namespace Azure.Communication.Sms
 
         /// <summary> Sends an SMS message from a phone number that belongs to the authenticated account. </summary>
         /// <param name="from"> The sender&apos;s phone number in E.164 format that is owned by the authenticated account. </param>
-        /// <param name="to"> The recipient&apos;s phone number in E.164 format. In this version, only one recipient in the list is supported. </param>
+        /// <param name="to"> The recipient&apos;s phone number in E.164 format. In this version, up to 100 recipients in the list is supported. </param>
         /// <param name="message"> The contents of the message that will be sent to the recipient. The allowable content is defined by RFC 5724. </param>
-        /// <param name="sendSmsOptions"> Optional configuration for sending SMS messages. </param>
+        /// <param name="options"> Optional configuration for sending SMS messages. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="RequestFailedException">The server returned an error. See <see cref="Exception.Message"/> for details returned from the server.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="from"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="to"/> is null.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="message"/> is null.</exception>
-        public virtual Response<SendSmsResponse> Send(PhoneNumberIdentifier from, IEnumerable<PhoneNumberIdentifier> to, string message, SendSmsOptions sendSmsOptions = null, CancellationToken cancellationToken = default)
+        public virtual Response<IEnumerable<SmsSendResult>> Send(string from, IEnumerable<string> to, string message, SmsSendOptions options = default, CancellationToken cancellationToken = default)
         {
             using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(SmsClient)}.{nameof(Send)}");
             scope.Start();
             try
             {
-                Argument.AssertNotNullOrEmpty(from.PhoneNumber, nameof(from));
-                return RestClient.Send(from.PhoneNumber, to.Select(x => AssertNotNullOrEmpty(x.PhoneNumber, nameof(to))), message, sendSmsOptions, cancellationToken);
+                Argument.AssertNotNullOrEmpty(from, nameof(from));
+                List<SmsRecipient> smsRecipients = new List<SmsRecipient>();
+                foreach (string toPhoneNumber in to)
+                {
+                    Argument.AssertNotNullOrEmpty(toPhoneNumber, nameof(to));
+                    SmsRecipient smsRecipient = new SmsRecipient(toPhoneNumber);
+                    smsRecipient.RepeatabilityRequestId = Guid.NewGuid().ToString();
+                    smsRecipient.RepeatabilityFirstSent = DateTime.Now.ToString(CultureInfo.InvariantCulture);
+                    smsRecipients.Add(smsRecipient);
+                }
+
+                Response<SmsSendResponse> response = RestClient.Send(from, smsRecipients, message, options, cancellationToken);
+                return Response.FromValue((IEnumerable<SmsSendResult>)response.Value.Value, response.GetRawResponse());
             }
             catch (Exception ex)
             {